tests/e2e/requirements.txt
pytest==8.3.0
httpx==0.27.0
influxdb-client==1.40.0
paho-mqtt>=1.6.1
